Rename ff_case
